perf: only call layout when it is needed
Make sure to only call `masonry.layout()` when something has changed for which it needs to layout. If we do not do this we are constantly applying the layout which results in DOM updates to the inline styles that can be avoided. This change also takes the height of the containing elements into account.